+++ CHANGES.md
@@ -20,6 +20,8 @@ UI:
 Storage:
 
 -   Add an API for storing and streaming content
+-   Add a DELETE endpoint
+-   Improves error handling (returns 404 from GET if the file doesn't exist)
 
 Others: